WSL: Prevent re-entrancy of stop()
This is the WSL equivalent of 0c0759752bb7d020e362a7806d56978f9a58d8ca:

stop() kills the subprocess, which can cause stop() to be called again.
This ends up confusing our progress code.  Prevent this with a flag.

+++ src/k8s-engine/wsl.ts
@@ -372,7 +372,14 @@ export default class WSLBackend extends events.EventEmitter implements K8s.Kuber
     }
   }
 
+  protected isStopping = false;
   async stop(): Promise<number> {
+    // When we manually call stop, the subprocess will terminate, which will
+    // cause stop to get called again.  Prevent the re-entrancy.
+    if (this.isStopping) {
+      return 0;
+    }
+    this.isStopping = true;
     try {
       this.setState(K8s.State.STOPPING);
       this.process?.kill('SIGTERM');
@@ -391,6 +398,8 @@ export default class WSLBackend extends events.EventEmitter implements K8s.Kuber
     } catch (ex) {
       this.setState(K8s.State.ERROR);
       throw ex;
+    } finally {
+      this.isStopping = false;
     }
 
     return 0;
